New cache strategy.

+++ sp-includes/CacheGenerator.php
@@ -7,22 +7,34 @@ class sp_UserCacheGenerator extends sp_FileReaderVisitor
 	private $conf_dir;
 	private $cache_dir;
 	private $meta_additional_fields;
-	private $routes;
+	private $last_update_stamp;
+	private $computed_update_stamp = 0;
 	
-	public function __construct(sp_StaticProjector $iSP)
+	private $current_file_stack = array();
+	private $current_file_order = null;
+	
+	public function __construct(sp_StaticProjector $iSP, $iUCStamp)
 	{
 		$this -> sp = $iSP;
 		$this -> basedir = $this -> sp -> basedir()."/".sp_StaticProjector::data_dir;
 		$this -> uc_dir = $this -> sp -> targetdir()."/".sp_StaticProjector::user_cache_dir;
 		$this -> conf_dir = $this -> sp -> targetdir()."/".sp_StaticProjector::config_dir;
 		$this -> cache_dir = $this -> sp -> targetdir()."/".sp_StaticProjector::cache_dir;
-		$this -> with_details = true;
+		$this -> with_details = false;
 		$this -> is_recursive = true;
 		$this -> meta_additional_fields = explode(";", sp_StaticProjector::file_metadata_additional_fields);
-		$this -> routes = array();
+		$this -> last_update_stamp = $iUCStamp;
 	}
 	
 	public function enter_directory(sp_FileInfo $info)
+	{
+		if(null != $this -> current_file_order)
+			array_push($this -> current_file_stack, $this -> current_file_order);
+		
+		$this -> current_file_order = array();
+	}
+	
+	public function exit_directory(sp_FileInfo $info)
 	{
 		$cache_dir = $this -> uc_dir.$info -> relative_path;
 		
@@ -30,46 +42,51 @@ class sp_UserCacheGenerator extends sp_FileReaderVisitor
 		if(!is_dir($cache_dir))
 		{
 			// Deleting any occurence of a directory which was a file before
-			if(file_exists($cache_dir))
-				unlink($cache_dir);
-
+			if(file_exists($cache_dir)) unlink($cache_dir);
 			@mkdir($cache_dir,null,true);
 		}
-
-		// Handling file order lists update
-		//{ // In fact PHP does not support anonymous blocks, lol
-			$local_vis = new sp_SimpleDirectoryContentList($info -> absolute_path);
-			$file_list = $local_vis -> get_list();
+		
+		$file_order_name = $cache_dir."/".sp_StaticProjector::file_order_name;
+		$local_stamp = max( file_exists($file_order_name) ? filemtime($file_order_name) : 0, $info -> timestamp_modified);
+		
+		if($this -> last_update_stamp < $local_stamp || ! file_exists($file_order_name))
+		{
+			$file_list = $this -> current_file_order;
 			$cache_list = array();
-			$file_order_name = $cache_dir."/".sp_StaticProjector::file_order_name;
-			
 			$previous_cache_list = array();
 			if(file_exists($file_order_name))
 			{
 				$cache_list = sp_ArrayUtils::load_config($file_order_name);
 				$previous_cache_list = $cache_list;
-				
+			
 				// Search the files which do not exists anymore
 				$not_existing_files = array_diff($cache_list,$file_list);
-
+			
 				// Delete those files from the cache
 				foreach ($not_existing_files as $file)
 				{
 					$path = $cache_dir."/$file";
 					if(file_exists($path))
 					{
-						$del = new sp_RecursiveDeleter($path);
-						$del->execute();
-
+						if(is_dir($path))
+						{
+							$del = new sp_RecursiveDeleter($path);
+							$del->execute();
+						}
+					}
+					
+					$meta = $path.sp_StaticProjector::file_metadata_ext;
+					if(file_exists($meta))
+					{
 						// Also delete rich data
-						$del = new sp_RecursiveDeleter($path.sp_StaticProjector::file_metadata_ext);
+						$del = new sp_RecursiveDeleter($meta);
 						$del->execute();
 					}
 				}
-
+			
 				// Remove them from the list
 				$cache_list = array_intersect($cache_list,$file_list);
-
+			
 				// Get the ones which are not in the list
 				$new_files = array_diff($file_list,$cache_list);
 				foreach ($new_files as $file)
@@ -81,86 +98,94 @@ class sp_UserCacheGenerator extends sp_FileReaderVisitor
 			{
 				$cache_list = $file_list;
 			}
-
+			
 			if($previous_cache_list != $cache_list)
-				sp_ArrayUtils::store_config($cache_list,$file_order_name);
-		//} // lol
+				sp_ArrayUtils::store_config(array_values($cache_list),$file_order_name);
+		}
 		
-		// Adding route patterns
-		if(empty($info -> relative_path))
-			array_push($this -> routes, "/ -> default()");
-		else
-			array_push($this -> routes, $info -> relative_path."/([^/]+) -> default()");
+		$this -> computed_update_stamp = max($this -> computed_update_stamp, filemtime($file_order_name));
+		$this -> current_file_order = array_pop($this -> current_file_stack);	
 	}
 	
 	public function process(sp_FileInfo $info)
 	{
-		$cache_file = $this -> uc_dir.$info -> relative_path;
-		// Deleting any occurence of a directory which is now a file
-		if(file_exists($cache_file) && is_dir($cache_file) && ! $info -> is_dir)
-		{
-			$del = new sp_RecursiveDeleter($cache_file);
-			$del->execute();
-		}
-		
-		// Adding rich data 
 		if(!empty($info -> relative_path))
 		{
+			$cache_file = $this -> uc_dir.$info -> relative_path;
 			$meta_file = $cache_file.sp_StaticProjector::file_metadata_ext;
-			$data = array();
 
-			if(file_exists($meta_file))
+			if($this -> last_update_stamp < $info -> timestamp_modified || ! file_exists($meta_file))
 			{
-				$data = sp_ArrayUtils::load_config($meta_file);
-			}
-			$previous_data = $data;
-			
-			if(!array_key_exists(sp_StaticProjector::file_metadata_title_field, $data))
-			{
-				$data[sp_StaticProjector::file_metadata_title_field] = $info->basename;
-			}
-			if(!array_key_exists(sp_StaticProjector::file_metadata_create_field, $data))
-			{
-				$data[sp_StaticProjector::file_metadata_create_field] = $info->last_modified_timestamp;
-			}
-			foreach($this->meta_additional_fields as $field)
-			{
-				if(!array_key_exists($field,$data))
+				// Deleting any occurence of a directory which is now a file
+				if(file_exists($cache_file) && is_dir($cache_file) && ! $info -> is_dir)
 				{
-					$data[$field]="";
+					$del = new sp_RecursiveDeleter($cache_file);
+					$del->execute();
+				}
+					
+				if($info -> is_dir)
+				{
+					$cache_dir = $this -> uc_dir.$info -> relative_path;
+					// Creating the folder if missing
+					if(!is_dir($cache_dir))
+					{
+						// Deleting any occurence of a directory which was a file before
+						if(file_exists($cache_dir)) unlink($cache_dir);
+						@mkdir($cache_dir,null,true);
+					}
+				}
+					
+				// Adding rich data
+				if(!empty($info -> relative_path))
+				{
+					$meta_file = $cache_file.sp_StaticProjector::file_metadata_ext;
+					$data = array();
+						
+					if(file_exists($meta_file))
+					{
+						$data = sp_ArrayUtils::load_config($meta_file);
+					}
+					$previous_data = $data;
+						
+					if(!array_key_exists(sp_StaticProjector::file_metadata_title_field, $data))
+					{
+						$data[sp_StaticProjector::file_metadata_title_field] = $info->basename;
+					}
+					if(!array_key_exists(sp_StaticProjector::file_metadata_create_field, $data))
+					{
+						//$data[sp_StaticProjector::file_metadata_create_field] = $info->timestamp_modified;
+						$data[sp_StaticProjector::file_metadata_create_field] = time();
+					}
+					foreach($this->meta_additional_fields as $field)
+					{
+						if(!array_key_exists($field,$data))
+						{
+							$data[$field]="";
+						}
+					}
+						
+					if($previous_data != $data)
+						sp_ArrayUtils::store_config($data, $meta_file);
 				}
 			}
-			
-			if($previous_data != $data)
-				sp_ArrayUtils::store_config($data, $meta_file);
+
+			$this -> computed_update_stamp = max($this -> computed_update_stamp, filemtime($meta_file));
+			array_push($this -> current_file_order, $info -> name);
 		}
-		
-		// Default route
-		if(!empty($info -> relative_path))
-			array_push($this -> routes, $info -> relative_path." -> default()");
 	}
 	
 	public function execute()
 	{	
 		if( ! $this -> is_processed())
 		{
-			if(! file_exists($this -> cache_dir))
-			{
-				@mkdir($this -> cache_dir, null, true);
-			}
+			$this -> computed_update_stamp = $this -> last_update_stamp;
 			parent::execute();
-			
-			$route_file = $this -> conf_dir."/".sp_StaticProjector::routes_file;
-			if(!file_exists($route_file))
-			{
-				touch($route_file);
-			}
-			if($this->sp->get_config()->default_routes_policy())
-				sp_ArrayUtils::dump_array($this -> routes, $this -> cache_dir."/".sp_StaticProjector::routes_default_file);
-			
-			sp_set_http_granting($this -> conf_dir, SP_HTTP_DENY_ACCESS);
-			sp_set_http_granting($this -> uc_dir, SP_HTTP_DENY_ACCESS);
 		}
+	}
+	
+	public function get_update_stamp()
+	{
+		return $this -> computed_update_stamp;
 	}
 }
 
@@ -180,6 +205,7 @@ class sp_PrivateCacheGenerator extends sp_FileReaderVisitor
 
 	private $meta_additional_fields;
 	private $debug = false;
+
 
 	public function __construct(sp_StaticProjector $iSP)
 	{
@@ -202,7 +228,7 @@ class sp_PrivateCacheGenerator extends sp_FileReaderVisitor
 		
 		$file_order_file = $this -> user_cache.$info -> relative_path.'/'.sp_StaticProjector::file_order_name;
 		sp_assert(file_exists($file_order_file));
-		$this -> current_file_order = file($file_order_file, FILE_IGNORE_NEW_LINES);
+		$this -> current_file_order = sp_ArrayUtils::load_config($file_order_file);
 	}
 	
 	public function exit_directory(sp_FileInfo $info)
@@ -237,18 +263,6 @@ class sp_PrivateCacheGenerator extends sp_FileReaderVisitor
 			sp_assert(is_dir($this -> cache_dir));
 			parent::execute();
 			sp_ArrayUtils::store_array($this -> dic_array, $this -> dic_file, $this -> debug);
-
-			// Parsing routes
-			$routes_data = array();
-			$routes = file($this -> sp -> targetdir().'/'.sp_StaticProjector::config_dir.'/'.sp_StaticProjector::routes_file, FILE_IGNORE_NEW_LINES);
-			foreach ($routes as $route_pattern)
-			{
-				if(preg_match('#^([^>\s]+)\s*->\s*([a-zA-Z0-9_\-]+)\s*\(([^\s]*)\)\s*$#',$route_pattern,$matches))
-				{
-					array_push($routes_data, array('route' => $matches[1], 'template' => $matches[2], 'replace_pattern' => $matches[3]));
-				}
-			}
-			sp_ArrayUtils::store_array($routes_data, $this -> sp -> targetdir().'/'.sp_StaticProjector::cache_dir.'/'.sp_StaticProjector::routes_dico, $this -> debug);
 		}
 	}
 }
@@ -257,27 +271,21 @@ class sp_PrivateCacheGenerator extends sp_FileReaderVisitor
 class sp_CacheGenerator 
 {
 	private $sp;
-	private $data_stamp = 0;
 	private $uc_stamp = 0;
 	private $cache_stamp = 0;
+	private $routes_stamp = 0;
 	
 	public function __construct(sp_StaticProjector $iSP)
 	{
 		$this -> sp = $iSP;
 	}
 	
-	public function check_current_filesystem_state()
-	{
-		$this -> data_stamp = sp_FileReader::get_directory_last_modified($this -> sp -> basedir().'/'.sp_StaticProjector::data_dir);
-		$this -> uc_stamp = sp_FileReader::get_directory_last_modified($this -> sp -> targetdir().'/'.sp_StaticProjector::user_cache_dir);
-		$this -> cache_stamp = sp_FileReader::get_directory_last_modified($this -> sp -> targetdir().'/'.sp_StaticProjector::cache_dir.'/'.sp_StaticProjector::dic_file);
-	}
-	
 	private function update_user_cache()
 	{		
-		$user_cache = new sp_UserCacheGenerator($this->sp);
+		$user_cache = new sp_UserCacheGenerator($this->sp, $this -> uc_stamp);
 		$user_cache -> execute();
 		$this -> sp -> log(sp_Logger::info, "Intermediate cache (".sp_StaticProjector::config_dir.") updated.");
+		$this -> uc_stamp = $user_cache -> get_update_stamp();
 	}
 	
 	private function generate_cache()
@@ -292,13 +300,46 @@ class sp_CacheGenerator
 		$cache_gen_policy = $this -> sp -> get_config() -> cache_policy();
 		if(sp_Config::cache_no_regen != $cache_gen_policy)
 		{
-			$update_user_cache = ($this -> uc_stamp <= $this -> data_stamp);
+			$stamp_data = array(0,0,0);
+			$cache_stamp_file = $this -> sp -> targetdir().'/'.sp_StaticProjector::cache_dir.'/'.sp_StaticProjector::cache_stamp_file;
+			if(file_exists($cache_stamp_file))
+			{
+				$stamp_data = sp_ArrayUtils::load_array($cache_stamp_file);
+			}
 
-			if($update_user_cache || sp_Config::cache_force_regen == $cache_gen_policy)
-				$this -> update_user_cache();
+			$this -> uc_stamp = $stamp_data[0];
+			$this -> cache_stamp = $stamp_data[1];
+			$this -> routes_stamp = $stamp_data[2];
+			
+			$this -> update_user_cache();
 
-			if( ($update_user_cache || $this -> cache_stamp <= $this -> uc_stamp) || sp_Config::cache_force_regen == $cache_gen_policy)
+			if($this -> cache_stamp < $this -> uc_stamp)
+			{
 				$this -> generate_cache();
+				$this -> cache_stamp = time();
+			}
+
+			$route_in = $this -> sp -> targetdir().'/'.sp_StaticProjector::config_dir.'/'.sp_StaticProjector::routes_file;
+			$route_out = $this -> sp -> targetdir().'/'.sp_StaticProjector::cache_dir.'/'.sp_StaticProjector::routes_dico;
+			$current_route_stamp = filemtime($route_in);
+			
+			if($this -> routes_stamp < $current_route_stamp)
+			{
+				// Parsing routes
+				$routes_data = array();
+				$routes = file($this -> sp -> targetdir().'/'.sp_StaticProjector::config_dir.'/'.sp_StaticProjector::routes_file, FILE_IGNORE_NEW_LINES);
+				foreach ($routes as $route_pattern)
+				{
+					if(preg_match('#^([^>\s]+)\s*->\s*([a-zA-Z0-9_\-]+)\s*\(([^\s]*)\)\s*$#',$route_pattern,$matches))
+					{
+						array_push($routes_data, array('route' => $matches[1], 'template' => $matches[2], 'replace_pattern' => $matches[3]));
+					}
+				}
+				sp_ArrayUtils::store_array($routes_data, $this -> sp -> targetdir().'/'.sp_StaticProjector::cache_dir.'/'.sp_StaticProjector::routes_dico);
+				$this -> routes_stamp = filemtime($route_out);
+			}
+			
+			sp_ArrayUtils::store_array(array($this -> uc_stamp, $this -> cache_stamp, $this -> routes_stamp), $cache_stamp_file);
 		}
 	}
 }
